fix problemas de assinaturas entre UseCase e Service

+++ src/main/java/br/com/uniceplac/sneaklab/application/ports/in/GerenciarPagamentoUseCase.java
@@ -8,17 +8,17 @@ import java.util.List;
 public interface GerenciarPagamentoUseCase {
 
     //Registra um novo pagamento para um pedido em RASCUNHO.
-    Pagamento registrarPagamento(int idPedido, double valor, TipoPagamento tipo);
+    Pagamento registrarPagamento(long idPedido, double valor, TipoPagamento tipo);
 
     //Aprova um pagamento PENDENTE e marca o pedido como PAGO.
-    Pagamento aprovarPagamento(int idPagamento);
+    Pagamento aprovarPagamento(long idPagamento);
 
     //Estorna um pagamento APROVADO e marca o pedido como CANCELADO.
-    Pagamento estornarPagamento(int idPagamento);
+    Pagamento estornarPagamento(long idPagamento);
 
     //Lista todos os pagamentos vinculados a um pedido.
-    List<Pagamento> listarPorPedido(int idPedido);
+    List<Pagamento> listarPorPedido(long idPedido);
 
     //Busca um pagamento específico pelo id.
-    Pagamento buscarPorId(int idPagamento);
+    Pagamento buscarPorId(long idPagamento);
 }
